Move server code into its folder within src
