+++ src/server/employees/emploees.service.spec.ts
@@ -1,10 +1,10 @@
 import { forwardRef } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { CompaniesModule } from 'src/companies/companies.module';
-import { CompaniesService } from 'src/companies/companies.service';
-import { Company } from 'src/companies/company.entity';
-import { OrdersModule } from 'src/orders/orders.module';
+import { CompaniesModule } from 'src/server/companies/companies.module';
+import { CompaniesService } from 'src/server/companies/companies.service';
+import { Company } from 'src/server/companies/company.entity';
+import { OrdersModule } from 'src/server/orders/orders.module';
 import { getConnection } from 'typeorm';
 import { Employee } from './employee.entity';
 import { EmployeesService } from './employees.service';
@@ -15,11 +15,11 @@ import {
   vouchersFactory,
   ordersFactory,
 } from 'test/factories';
-import { PartnersModule } from 'src/partners/partners.module';
-import { VouchersModule } from 'src/vouchers/vouchers.module';
-import { VouchersService } from 'src/vouchers/vouchers.service';
-import { PartnersService } from 'src/partners/partners.service';
-import { OrdersService } from 'src/orders/orders.service';
+import { PartnersModule } from 'src/server/partners/partners.module';
+import { VouchersModule } from 'src/server/vouchers/vouchers.module';
+import { VouchersService } from 'src/server/vouchers/vouchers.service';
+import { PartnersService } from 'src/server/partners/partners.service';
+import { OrdersService } from 'src/server/orders/orders.service';
 
 describe('EmployeesService', () => {
   let employeesService: EmployeesService;
