+++ src/server/employees/employees.module.ts
@@ -1,8 +1,8 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { CompaniesModule } from 'src/companies/companies.module';
-import { Employee } from 'src/employees/employee.entity';
-import { OrdersModule } from 'src/orders/orders.module';
+import { CompaniesModule } from 'src/server/companies/companies.module';
+import { Employee } from 'src/server/employees/employee.entity';
+import { OrdersModule } from 'src/server/orders/orders.module';
 import { EmployeesResolver } from './employees.resolver';
 import { EmployeesService } from './employees.service';
