+++ src/server/orders/orders.module.ts
@@ -1,7 +1,7 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { EmployeesModule } from 'src/employees/employees.module';
-import { VouchersModule } from 'src/vouchers/vouchers.module';
+import { EmployeesModule } from 'src/server/employees/employees.module';
+import { VouchersModule } from 'src/server/vouchers/vouchers.module';
 import { Order } from './order.entity';
 import { OrdersResolver } from './orders.resolver';
 import { OrdersService } from './orders.service';
