+++ src/server/partners/partners.module.ts
@@ -1,7 +1,7 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { OrdersModule } from 'src/orders/orders.module';
-import { VouchersModule } from 'src/vouchers/vouchers.module';
+import { OrdersModule } from 'src/server/orders/orders.module';
+import { VouchersModule } from 'src/server/vouchers/vouchers.module';
 import { Partner } from './partner.entity';
 import { PartnersResolver } from './partners.resolver';
 import { PartnersService } from './partners.service';
