+++ src/server/partners/partners.resolver.ts
@@ -2,7 +2,7 @@ import { Resolver, Query, ResolveField, Parent } from '@nestjs/graphql';
 import { Inject } from '@nestjs/common';
 import { Partner } from './partner.entity';
 import { PartnersService } from './partners.service';
-import { VouchersService } from 'src/vouchers/vouchers.service';
+import { VouchersService } from 'src/server/vouchers/vouchers.service';
 
 @Resolver((_of) => Partner)
 export class PartnersResolver {
