+++ src/server/schema.gql
@@ -1,0 +1,75 @@
+# ------------------------------------------------------
+# THIS FILE WAS AUTOMATICALLY GENERATED (DO NOT MODIFY)
+# ------------------------------------------------------
+
+type Partner {
+  id: Float!
+  name: String!
+  vouchers: [Voucher]!
+  created_at: DateTime!
+  updated_at: DateTime!
+  revenue: Float!
+}
+
+"""
+A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format.
+"""
+scalar DateTime
+
+type Voucher {
+  id: Float!
+  amount: Float!
+  partner: Partner!
+  orders: [Order]!
+  created_at: DateTime!
+  updated_at: DateTime!
+}
+
+type Order {
+  id: Float!
+  date: DateTime!
+  voucher: Voucher!
+  employee: Employee!
+  created_at: DateTime!
+  updated_at: DateTime!
+}
+
+type Employee {
+  id: Float!
+  name: String!
+  budget: Float!
+  company: Company!
+  orders: [Order]!
+  created_at: DateTime!
+  updated_at: DateTime!
+  spend: Float!
+  spendInMonth(month: DateTime!): SpendDto!
+}
+
+type Company {
+  id: Float!
+  name: String!
+  employees: [Employee]!
+  created_at: DateTime!
+  updated_at: DateTime!
+}
+
+type TaxableSpend {
+  thirtyPercentBracket: Float!
+}
+
+type SpendDto {
+  total: Float!
+  taxFree: Float!
+  taxable: TaxableSpend!
+}
+
+type Query {
+  employees: [Employee!]!
+  employeesByCompany(companyId: Int!): [Employee!]!
+  companies: [Company!]!
+  orders: [Order!]!
+  order(id: Int!): Order!
+  vouchers: [Voucher!]!
+  partners: [Partner!]!
+}
